Maintain dynamic plugin metadata inside module
For each module, create dynamic-plugs.ts file that stores both
'exposedModules'(as defined in plugin.json) and 'extensions'(as defined
in console-extension.ts).
