--- src/extensions/HostsPageWrapper.tsx
@@ -1,7 +1,0 @@
-import { HostsPage } from '@app/Providers/HostsPage';
-
-import withQueryClient from './QueryClientHoc';
-
-const Page = withQueryClient(HostsPage);
-
-export default Page;
