--- src/extensions/MappingsWrapper.tsx
@@ -1,7 +1,0 @@
-import { MappingsPage } from '@app/Mappings/MappingsPage';
-
-import withQueryClient from './QueryClientHoc';
-
-const Page = withQueryClient(MappingsPage);
-
-export default Page;
