--- src/extensions/PlanWizardWrapper.tsx
@@ -1,7 +1,0 @@
-import { PlanWizard } from '@app/Plans/components/Wizard/PlanWizard';
-
-import withQueryClient from './QueryClientHoc';
-
-const Page = withQueryClient(PlanWizard);
-
-export default Page;
