--- src/extensions/ProvidersWrapper.tsx
@@ -1,7 +1,0 @@
-import { ProvidersPage } from '@app/Providers/ProvidersPage';
-
-import withQueryClient from './QueryClientHoc';
-
-const Page = withQueryClient(ProvidersPage);
-
-export default Page;
