+++ src/modules/Hosts/HostsPageWrapper.tsx
@@ -1,0 +1,7 @@
+import withQueryClient from 'src/components/QueryClientHoc';
+
+import { HostsPage } from '@app/Providers/HostsPage';
+
+const Page = withQueryClient(HostsPage);
+
+export default Page;
