+++ src/modules/Plans/PlanWizardWrapper.tsx
@@ -1,0 +1,7 @@
+import withQueryClient from 'src/components/QueryClientHoc';
+
+import { PlanWizard } from '@app/Plans/components/Wizard/PlanWizard';
+
+const Page = withQueryClient(PlanWizard);
+
+export default Page;
